requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-spectacular==0.24.2
